requirements.txt
openai==1.7.2
python_telegram_bot==20.7
PyYAML==6.0.1
telegram==0.0.1
